refactor utils and helpers file names
