+++ src/js/controllers/mapController.ts
@@ -44,7 +44,7 @@ import { addPopupWatchUtils } from 'js/helpers/DataPanel';
 
 import { createAndAddNewGraphic } from 'js/helpers/MapGraphics';
 
-import { getCustomSymbol } from 'js/utils/symbol.config';
+import { getCustomSymbol } from 'js/helpers/generateSymbol';
 
 const allowedLayers = ['feature', 'dynamic', 'loss', 'gain']; //To be: tiled, webtiled, image, dynamic, feature, graphic, and custom (loss, gain, glad, etc)
