+++ src/js/helpers/coordinatesConversion.ts
