+++ src/js/helpers/generateSymbol.ts
