+++ src/js/helpers/spatialDataTransformation.ts
